Merged PR 8263219: Enable host based tests for DxeRust
Reworks a bug in the CI plugin - Previously it would compile binaries as tests too, which does not work for binaries compiled as EFI drivers as they are target specific. Updates the plugin to only compile the library (and tests associated with the library) and integration tests found in the tests/* folder.

Adds some conditional compilation to DxeRust package so that the library can be compiled on host systems.