Merged PR 8607003: Simplify rust unit test plugin
Rename CargoTestHostCheck to HostBasedRustUnitTestRunner.

Greatly simplifies the plugin due to being able to use cargo make
to run the tests.

Additionally adds a IUefiHelperPlugin that registers a RustPackage class, allowing this class to be passed throughout the build system. Currently it is only used for the HostBasedRustUnitTestRunner plugin but it will soon be used by the RustCoverageCheck plugin.