Merged PR 8953277: Makefile.toml: multiple bugfixes
Adds support for building with the local arch with `-e ARCH=LOCAL`

bugfix an issue not allowing the user to override `TARGET_TRIPLE` directly. This bug was due to the fact that the ENV setting would only execute if TARGET_TRIPLE was not set, however it was always being due to the use of `default_value`.

simplifies the `cargo make build` command such that the underlying cargo command is `cargo build` rather then `cargo rustc`.

## Integration Instructions
If pulling in the commit and submodule without a fresh clone, delete your Conf folder (or run stuart_build with the `--UPDATECONF` flag