Merged PR 8988833: Readme: Update cargo make build information
Updates the `cargo make build` command information in the readme to
inform users that they can now build a package's binaries using the
local architecture (ARCH=LOCAL), which is useful for library packages
that contain locally executable binary tools (such as a depex parser).

Updates the `cargo make build` command information in the readme to
inform users that they must now specify a package to build instead of
being able to execute `cargo make build` to build all packages. This
is due to the fact that, with the addition of locally compilable tools,
it is no longer true that all packages's are able to compile to
x86_64-unknown-uefi|i686-unknown-uefi.