DxeRust/fv: Do not return pad files in get_next_files()
Per the PI Spec (PI Spec 1.8 - 2.1.4.1.8 EFI_FV_FILETYPE_FFS_PAD)
pad files should not be returned from file system services:

- https://uefi.org/specs/PI/1.8/V3_Design_Discussion.html#efi-fv-filetype-ffs-pad

Doing so causes an assert in RustFfiTestDxe to occassionally trigger
because the file data size returned from `Fv->GetNextFile()` is
`0`.

This change does not return an unknown or pad file to the caller
that might be encountered.