Merged PR 9306680: Rework UefiProtoclDb implementation to use a Map rather than a Vec for storing handles.
Rework UefiProtoclDb implementation to use a Map rather than a Vec for storing handles. This has a number of advantages:
1. No longer need a bunch of `-1` arithmetic to convert handles to vector indexes.
2. Deleting all protocols on a handle allows us to remove it from the map, thus actually reclaiming the space; in the vec implementation the handle slot would still be in the vec consuming memory, but would become inaccessible.

This implementation may be slightly less performant since it requires a map.get() operation rather than direct vector indexing for most handle operations. My expectation is that this will be negligible in practice as the number of handles is typically not large (I expect it to be in the low hundreds).