Merged PR 9339927: Add implementation of GCD memory services
# PR to add memory services in GCD

`sorted_slice.rs` : this file contains an implementation for a struct that offers functionality to add items and keep them sorted in a given buffer.

`memory_block.rs` : This file contains an implementation for the structure of a block of memory stored by the GCD. This file also contains the state transition of the memory block.

`gcd.rs` : Put together the previous file to offer the memory service (add, remove, allocate, free). This file also contains all the tests to validate the behavior of the gcd.