Merged PR 9648547: Switch to using PEI/DxeIpl-created page tables
Switch to using PEI/DxeIpl-created page tables. The x86_64 created page tables are not complete - they are missing pages corresponding to high MMIO, for example. This causes issues with CPU arch driver later on.

This PR removes them until we either decide to simply roll with the DxeIpl-created tables, or update the page table implementation we have to be faster and more complete.