Merged PR 9897765: Fix an issue where Loaded Image Device Path pointed to invalid memory.
The loaded image device path installed on the image handle could point to invalid memory when specified by caller. The previous implementation was simply exposing the caller-supplied pointer to device path as the device path protocol on the image handle; however caller was not required to ensure that pointer was long lived. The correct behavior (which mirrors the C reference implementation) is to copy the device path supplied by the caller into a newly allocated buffer so that it will be permanent.

The invalid device path was causing hangs/long processing when the 'dh' command in shell attempted to decode it.

This PR adds copy_device_path_to_boxed_slice and uses it to ensure ownership of loaded image device path.