Merged PR 9933139: Fix a Tpl Lock Assertion on exit from shell
Fix an issue where start_image was incorrectly holding the private image context. This was introduced inadvertently as part of start_image factor to resolve clippy issues in !9927597.

The issue was caused by start_image() holding the lock on PRIVATE_IMAGE_DATA during the call to core_unload_image(). The change moves the PRIVATE_IMAGE_DATA lock inside appropriate scopes to ensure that it is released before the call to core_unload_image().