Merged PR 9956554: Fix HII image extraction on Linux builds
The previous HII image extraction method attempted to return the HII
data as a slice within the resource section. This worked on resource
sections constructed in Windows but the offsets vary due to slight
differences on Linux built resources. This follows the C
implementation's approach to base the resource section using the
data entry offset relative to the loaded image base which works for
both Linux and Windows.