Merged PR 10027335: Add ability to deallocate pages from GCD
Currently page allocations come from the same infrastructure as pool
allocations. This change updates page allocations to come directly
from the GCD. This means a new allocator is created on a page
allocation to track it and the allocation node it designated as
directly from the GCD so during deallocation, the page allocation
can be freed from the GCD.

In particular, this is necessary for OS allocation patterns that
allocate EfiLoader pages from any address, free the pages, and then
allocate a subset of pages from the same address.