DxeRust/protocols.rs: Limit println!() to non-test
The SerialPrintDxe crate is used for output. It is being updated to
support an "std" feature flag. In the meantime, the no_std _print()
function is executing causing the following failures for any test
code paths through it:

```
Caused by:
  process didn't exit successfully: `C:\src\UefiRust\target\debug\deps
  \dxe_rust-732c5af612c9e728.exe --nocapture` (exit code: 0xc0000096,
  STATUS_PRIVILEGED_INSTRUCTION)
```

So, SerialPrintDxe is being limited here to non-test scenarios to
unblock the unit tests being added.

protocols.rs fixup