Merged PR 10612998: Use binary Rust DXE Core
Pulls in the Rust DXE Core as a binary via an ext dep. This allows
the DXE Core to be shared among several repos with consistent
versioning and tracking, focuses development on the DXE Core in its
own repo, and improves separation between the DXE Core and its
support/dependencies from that of platforms that integrate it.