Merged PR 11425458: Default SHUTDOWN_AFTER_RUN to FALSE
Default SHUTDOWN_AFTER_RUN to FALSE

SHUTDOWN_AFTER_RUN is a build flag that is passed to QemuRunner to
instruct it to close the QEMU instance after startup.nsh has been
run. This is useful for automated testing. It is however annoying
for local testing, as one must either catch the EFI shell in the
5 second window it counts down before running the script or remember
to put SHUTDOWN_AFTER_RUN=FALSE on the cmdline.

mu_tiano_platforms defaults SHUTDOWN_AFTER_RUN to FALSE, as that is
the expected local setup and the CI scripts specify it to be TRUE.
This PR moves this repo to the same state. The CI scripts already
specify SHUTDOWN_AFTER_RUN=TRUE.

----
#### AI description  (iteration 1)
#### PR Classification
Configuration change to modify default behavior.

#### PR Summary
This pull request changes the default value of the `SHUTDOWN_AFTER_RUN` configuration to `FALSE`.
- `Platforms/QemuQ35Pkg/PlatformBuild.py`: Set `SHUTDOWN_AFTER_RUN` default to `FALSE`.
- `Platforms/QemuQ35Pkg/Plugins/QemuRunner/QemuRunner.py`: Updated to reflect the new default value for `SHUTDOWN_AFTER_RUN`.