Merged PR 11374185: PlatformBuild.py: Force full build if BLD_*_DXE_CORE_BINARY_PATH set
PlatformBuild.py: Force full build if BLD_*_DXE_CORE_BINARY_PATH set

The pipelines in [DxeRust](https://dev.azure.com/microsoft/MsUEFI/_git/DxeRust)
attempt to build and run UefiRust (QemuQ35Pkg and QemuSbsaPkg) as a part
of Continuous Integration, however `stuart_pr_eval` is causing the
builds to be skipped as there is technically no change in any file in
UefiRust. This PR adds a check to see if the
`BLD_*_DXE_CORE_BINARY_PATH` is defined on the command line. If it is
present, it will return that the platform must be built.

This change is necessary because we use the pipeline templates from
`mu_devops`, which cannot turn off `stuart_pr_eval`.`

----
#### AI description  (iteration 1)
#### PR Classification
Code enhancement to ensure a full build when specific environment variables are set.

#### PR Summary
This pull request modifies the build scripts to force a full build if the `BLD_*_DXE_CORE_BINARY_PATH` environment variable is set.
- `Platforms/QemuQ35Pkg/PlatformBuild.py`: Added a check to return all potential packages for building if the `BLD_*_DXE_CORE_BINARY_PATH` variable is set.
- `Platforms/QemuSbsaPkg/PlatformBuild.py`: Added a similar check to ensure all potential packages are built if the `BLD_*_DXE_CORE_BINARY_PATH` variable is set.