Merged PR 11660585: Use the qemu_rust_bins DXE Core NuGet binary
Replaces the Rust DXE Core NuGet package used before with the package
produced from qemu_rust_bins (DXECORE.QEMU).

The logic in the Python files for adjusting the path variable based
on whether a DEBUG or RELEASE build is used is removed for a simpler
approach of specifying the .efi binary path in the FDF based on the
target type.

The Rust DXE Core binary can be overridden on the command line as
follows by setting the `BLD_*_DXE_CORE_BINARY_PATH` to the location
for the target build.

For example:

- **Q35**: Add `BLD_*_DXE_CORE_BINARY_PATH=C:\src\qemu_rust_bins\target\x86_64-unknown-uefi`
- **SBSA**: Add `BLD_*_DXE_CORE_BINARY_PATH=C:\src\qemu_rust_bins\target\aarch64-unknown-uefi`