Merged PR 12422521: Build, patch, run script: Always copy code FD files
The script originally made a copy of the code FD file with the
extension ".ref.fd" that persisted until deleted by the user.
The actual code file was patched. The idea being that after a new
build, the user would manually remove the reference file to acknowledge
they are ready to move to a new reference.

To simplify usage, this change always patches the current code file
so it will pick up build changes outside of patching automatically.

The ".ref.fd" file is deleted after it is used for the patch step.

The black code formatter is also run against the file.