GitHub CI Transition: Add Platform CI GitHub workflow
Converts the CodeQL GitHub workflow to be used for generic platform
CI in this repository. It is intentionally limited in what it checks
to reduce complexity and maintenance in this repo. Right now, it
only supports QemuQ35Pkg build with VS2022. It should be updated to
support QemuSbsaPkg build soon.

In future merges from the upstream, this file should be left alone
unless specific changes are required in it.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>