Add FF-A for SBSA platform (#1150)
This change introduces the ARM FF-A feature into SBSA platform.

A few features are still in the process of being upstreamed to their
mainline. But here is a functional version that boots with Windows.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [ ] Impacts security?
- [x] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

This change was tested by booting to Windows on virtual platforms.

Developers need to install newly required toolchain using:
- `sudo apt-get install -y libssl-dev clang llvm lld
device-tree-compiler`
- `pip install -r requirements.txt`

---------

Signed-off-by: kuqin12 <42554914+kuqin12@users.noreply.github.com>
Co-authored-by: Raymond-MS <raymonddiaz@microsoft.com>
Co-authored-by: Chris Fernald <chfernal@microsoft.com>