Use BLD_*_PERF_TRACE_ENABLE=TRUE to enable performance measurement. (#15)
## Description

Use `BLD_*_PERF_TRACE_ENABLE=TRUE` to enable performance measurement.
Performance enabled by default was a change that was introduced at the
time because performance was always enabled in DXE core. Now that
performance is a component that can be added and removed, the original
way of setting it up is preferred. If the performance component is added
to a build, `BLD_*_PERF_TRACE_ENABLE` should be true, otherwise false.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Run on QEMU.

## Integration Instructions

N/A