QemuSbsaPkg/PlatformPeiLib: Remove FV HOB install
This installs a FV HOB that describes the entire QEMU_EFI.FD image.

This image contains a compressed section for FVMAIN_COMPACT.Fv that
is decompressed in PEI and a FV HOB is installed of the decompressed
image. The Patina DXE Core gets a FV HOB of the uncompressed image
which it immediately operates on to begin dispatching images and
then extracts the compressed section in the overall QEMU_EFI.FD HOB
which leads to everything being decompressed and duplicate copies
of drivers in memory and being dispatched. This change removes the
FV HOB installation for the uncompressed FV.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>