build_and_run_rust_binary.py: Linux QoL improvements (#52)
## Description

The script was originally written assuming QemuSbsaPkg would primarily
be built using a CLANGPDB environment on Windows. This change makes some
tweaks for it to operate better in a Linux and GCC environment.

1. Set the patch repo default to a workspace in the parent directory of
the patina-qemu directory. This works on both Windows and Linux.
2. Set the SBSA toolchain default to GCC5 instead of CLANGPDB.
3. Set the backup QEMU directory to the system bin directory instead of
the Windows QEMU ext dep.

There is also a small formatter change included from black formatter.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

- `build_and_run_rust_build.py` on both Windows (VS2022/Q35) and Ubuntu
(GCC5/SBSA)

## Integration Instructions

- N/A

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>