Update platform-ci to target more build configs
This commit updates the platform-ci workflow to remove all unnecessary
setup steps for building rust based UEFI drivers, which is no longer
needed. It also updates the build matrix to include more build
configurations, and to produce the repository constants through a
different workflow, so that they can be reused in other workflows.