[CHERRY-PICK] Disable SVE/SME on QemuSbsaPkg (#1233) (#64)
## Description

Cherry-picked from
https://github.com/microsoft/mu_tiano_platforms/commit/e28742f0fad5e12d71ff649a3866febe9f5ae03c.

When FF-A support was added to QemuSbsaPkg, it re-enabled SVE and SME
for TF-A and QEMU, which had been disabled because Windows boot was
broken with those.

Windows boot is again broken because of SVE/SME and a [change in
ArmPlatformPkg](https://github.com/microsoft/mu_silicon_arm_tiano/commit/cb87aada970c68c1a210ed68a4a1ce238623e3c3)
consumed from edk2 that aligned the feature register reserved field
behavior with the spec. A Windows change will be forthcoming, but in the
mean time, SVE/SME needs to be disabled again. The edk2 change is
technically correct, so that change is left in place.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Booting SBSA to Windows was failing; now it is succeeding.

## Integration Instructions

Rebuild TF-A with the PlatformBuild.py changes, boot qemu with the
changes in QemuRunner.py.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>