Fix Building Release patina-dxe-core-qemu
The current build and run script was incorrectly trying to
just pass --profile release to initiate a release build in
patina-dxe-core-qemu, however this does not work with how
patina-dxe-core-qemu's Makefile.toml is set up (and has been
for some time).

This commit fixes the script to call cargo make [q35|sbsa]-release
when doing a release build.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>