Build and Run Script: Update Default Qemu Path Per Arch
Currently, when running the build and run script for Q35,
the Windows path is always the default QEMU path. This changes
the script to use the Windows path if on Windows and the common
Linux path (what the container uses) for Linux by default. It
can still be explicitly overridden via CLI.

SBSA is not changed and the Linux path is left as the default
because at this point it must be built in Linux.